% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_variable.R
\name{custom_variable}
\alias{custom_variable}
\title{Add Custom Variable.}
\usage{
custom_variable(data, custom_var, level, data_type, label = custom_var,
  values = NULL, config_csv = NULL)
}
\arguments{
\item{data}{List of data.table obejcets returned by read_data().}

\item{custom_var}{Character. New custom variable name.}

\item{level}{Either "household", "person", "vehicle", or "trip".}

\item{label}{Description of custom variable. Defaults to custom_var value.}

\item{values}{Values of the custom variable. Must be same length as destination table. Not needed when using config_csv.}

\item{config_csv}{File path to a csv with fields "DOMAIN", "VALUE", "DESCRIPTION". Not needed when using values.}
}
\description{
Add custom variable to dataset and codebook.
}
