% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{derived_variables}
\alias{derived_variables}
\title{Add derived Variables from a configuration worksheet (csv).}
\usage{
derived_variables(data, config_csv = NULL)
}
\arguments{
\item{data}{Object returned by \link[summarizeNHTS]{read_data}.}

\item{config_csv}{File path to a csv with fields "NAME", "TABLE", "TYPE", "DOMAIN", "VALUE", "LABEL".}
}
\description{
Add custom variable to dataset and codebook.
}
