% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nhts_data.R
\name{download_nhts_data}
\alias{download_nhts_data}
\title{Download NHTS datasets.}
\usage{
download_nhts_data(dataset, exdir = getwd())
}
\arguments{
\item{dataset}{The study year of the dataset to download. Currently either '2001', '2009', or '2017'.}

\item{exdir}{The directory to house the data. Defaults to working directory.}
}
\description{
Download public use NHTS datasets from the web.
}
\details{
We recommend that you first create an \href{https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects}{RStudio Project}.
This will ensure that the data is downloaded under your project directory and will be helpful for future package utilization.
Note: A "/csv/dataset/" directory will always be created under exdir for organizational purposes.
}
\examples{
\donttest{
# Download the 2009 NHTS data to specified directory:
download_nhts_data("2009", "C:/NHTS")
}

}
