% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bar_chart.R
\name{make_bar_chart}
\alias{make_bar_chart}
\title{Make Interactive Bar Charts}
\usage{
make_bar_chart(tbl, facet = F, order = F, color_palette = "Set1",
  interactive = T, flat_print = F, squeeze_coord = "x", flip_coord = F,
  digits = 2, percentage = attr(tbl, "prop"), scientific = F,
  multiplier = NULL)
}
\arguments{
\item{tbl}{data.table object returned by \link[summarizeNHTS]{make_table}}

\item{facet}{logical. Whether to call \link[ggplot]{facet_grid} when using more than one variable}

\item{order}{logical. Order by values instead of natural codebook order?}

\item{interactive}{logical. Print ggiraph object without interactivity?}

\item{flat_print}{logical. Print as ggpplot object instead ggiraph object?}

\item{squeeze_coord}{character. "x" or "y". The default ("x") means the variable with the most groups will be the bar chart's x variable}

\item{flip_coord}{logical. Flip x and y axes?}

\item{digits}{integer. Number of significant digits to use.}

\item{percentage}{logical. Treat proportions as percentages?}

\item{scientific}{logical. Use scientific notation for number formatting?}

\item{multiplier}{numeric. A multiplier to use for numeric value display (i.e. For "In Thousands", use multiplier = 1000)}

\item{color_pallete}{A color pallete. See \link[RColorBrewer]{display.brewer.all}}
}
\description{
Make interactive bar charts utilizing \link[ggiraph]{ggiraph}.
}
