% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chart.R
\name{make_chart}
\alias{make_chart}
\title{Make Interactive Charts}
\usage{
make_chart(tbl, x = NULL, y = NULL, fill = NULL, facet = NULL,
  interactive = TRUE, order = FALSE, flip = FALSE, flat_print = FALSE,
  palette = "Set1", digits = 2, percentage = attr(tbl, "prop"),
  scientific = F, multiplier = NULL, ...)
}
\arguments{
\item{tbl}{data.table object returned by \link[summarizeNHTS]{make_table}.}

\item{x}{character. Variable to be displayed along x-axis.}

\item{y}{character. Either 'W' for weighted or 'U' for unweighted statistic.}

\item{fill}{character. Variable to be categorized using interior coloring.}

\item{facet}{logical. Variable to be categorized by faceted grid.}

\item{interactive}{logical. Print as ggpplot object instead ggiraph object?}

\item{order}{logical. Order by values instead of natural codebook order?}

\item{flip}{logical. Flip x and y axes?}

\item{flat_print}{logical. Print as ggpplot object instead ggiraph object?}

\item{palette}{character. A color pallete to be used for the fill variable. See \link[RColorBrewer]{display.brewer.all}}

\item{digits}{integer. Number of significant digits to use.}

\item{percentage}{logical. Treat proportions as percentages?}

\item{scientific}{logical. Use scientific notation for number formatting?}

\item{multiplier}{numeric. A multiplier to use for numeric value display (i.e. For "In Thousands", use multiplier = 1000)}
}
\description{
Make interactive bar charts utilizing \link[ggiraph]{ggiraph}.
}
