% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html_table.R
\name{make_html_table}
\alias{make_html_table}
\title{Make HTML Table}
\usage{
make_html_table(ftbl, title = "")
}
\arguments{
\item{ftbl}{An ftable object returned by \link[summarizeNHTS]{make_crosstab}}

\item{title}{Title of the table}
}
\description{
Make formatted HTML tables utilizing \link[htmlTable]{htmlTable}
}
