% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create weighted aggregate tables using NHTS data.}
\usage{
make_table(data, agg, agg_var = NULL, factors = NULL, wgt_name,
  variance = "se", subset = TRUE, jk_coeff = 99/100, var_lookup = if
  (exists("variables")) variables else NULL)
}
\arguments{
\item{data}{data.table object containing relavent NHTS variables and weights}

\item{agg}{Aggregate function label ("count","sum","avg", or "avg_trip_rate")}

\item{agg_var}{Variable name to aggregate over. Only relavent when agg is "avg" or "sum"}

\item{factors}{Character element or vector of variable names to group by}

\item{wgt_name}{Name of the weight variable being used}

\item{variance}{Variance calculation to be used. Either "se" for Standard Error or "moe" for Margin of Error}

\item{subset}{A Pre-aggregation subset condition}

\item{jk_coeff}{Jacknife coefficient for standard error calculations}
}
\value{
data.table object aggregated by input specifications
}
\description{
Create weighted aggregate tables using NHTS data.
}
\examples{
# Example 1: Let's get the aveage number of vehicle miles by state and derived familiy income.

# Create vector of weight replicate names
wgt_names <- get_wgt_names('DAYWGT')

# Read in the two necessary datasets
per50wt <- fread('./data/2009/per50wt.csv', select = c('HOUSEID','PERSONID',wgt_names))
DAYV2PUB <- fread('./data/2009/DAYV2PUB.csv',select = c('HOUSEID','PERSONID','HHFAMINC','VMT_MILE','HHSTATE'))

# Set the appropriate unique identifiers for merging
setkeyv(per50wt, c('HOUSEID','PERSONID'))
setkeyv(DAYV2PUB, c('HOUSEID','PERSONID'))

# merge the datasets
dt <- DAYV2PUB[per50wt, nomatch=0]

ex3 <- make_table(
  data = dt, 
  agg = 'avg',
  agg_var = 'VMT_MILE',
  factors = c('HHSTATE','HHFAMINC'),
  wgt_name = 'DAYWGT',
  subset = 'VMT_MILE >= 0',
  variance = 'moe' #Option to get Margin of Error, instead of default Standard Error
)
}
