% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhts_2009}
\alias{nhts_2009}
\title{List including the 2009 NHTS codebook labels and variables.}
\format{A list with 2 elements
\itemize{
  \item labels. Contains variable name, answer code, and label.
  \item variables. Contains variable name, description, datafile, question type, data type, size, and level.
}}
\description{
List including the 2009 NHTS codebook labels and variables.
}
\keyword{data}
