% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read HTS data.}
\usage{
read_data(dataset, select = select_all(dataset), csv_path = getwd())
}
\arguments{
\item{dataset}{The study year of the dataset.}

\item{select}{A character vector of NHTS variable names to select for analysis. Defaults to all.}

\item{csv_path}{The parent directory of "/csv/dataset/". Defaults to working directory.}
}
\description{
Read and merge NHTS data for analysis.
}
\details{
\code{read_data} is a wrapper for reading in variables from the correct csvs
and merging the corresponding tables on correct keys. \link[data.table]{fread} and
\link[data.table]{merge.data.table} are used for performance benefits.
}
