% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhts_data.R
\name{read_nhts_data}
\alias{read_nhts_data}
\title{Read NHTS data.}
\usage{
read_nhts_data(dataset, select = select_all(dataset), csv_path = getwd())
}
\arguments{
\item{dataset}{The study year of the dataset to read Currently only '2009'.}

\item{select}{A character vector of NHTS variable names to select for analysis. 
See \link[summarizeNHTS]{nhts_2009} for more info.}

\item{csv_path}{The parent directory of "/csv/dataset/". Defaults to working directory.}
}
\description{
Read and merge NHTS data for analysis.
}
\details{
\code{read_nhts_data} is a wrapper for reading in variables from the correct csvs
and merging the corresponding tables on correct keys. \link[data.table]{fread} and
\link[data.table]{merge.data.table} are used for performance benefits.
}
