% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chart.R
\name{make_chart}
\alias{make_chart}
\title{Make Interactive Charts}
\usage{
make_chart(tbl, x = NULL, y = NULL, fill = NULL, facet = NULL,
  interactive = TRUE, order = FALSE, flip = FALSE,
  flat_print = FALSE, legend = TRUE, variable_labels = TRUE,
  palette = "Set1", confidence = 0.95, label_options = list(wrap =
  35, trunc = 100), ggiraph_options = list(), ...)
}
\arguments{
\item{tbl}{data.table object returned by \link[summarizeNHTS]{summarize_data}.}

\item{x}{character. Variable to be displayed along x-axis.}

\item{y}{character. Either 'W' for weighted or 'U' for unweighted statistic.}

\item{fill}{character. Variable to be categorized using interior coloring.}

\item{facet}{logical. Variable to be categorized by faceted grid.}

\item{interactive}{logical. Print as ggpplot object instead ggiraph object?}

\item{order}{logical. Order by values instead of natural codebook order?}

\item{flip}{logical. Flip x and y axes?}

\item{flat_print}{logical. Print as ggpplot object instead ggiraph object?}

\item{legend}{logical. Display a legend?}

\item{variable_labels}{logical. Use variable labels/descriptions?}

\item{palette}{character. A color pallete to be used for the fill variable. See pallete argument here \link[ggplot2:scale_brewer]{scale_fill_brewer}.}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.95. Set to NULL for standard error.}

\item{label_options}{List of label options when variable_labels = TRUE. Number of characters before wrapping and truncating text, respectively.}

\item{ggiraph_options}{Optional list of options to pass to \link[ggiraph]{ggiraph}.}

\item{...}{Optional formatting arguments. See \link[summarizeNHTS]{format_values}.}
}
\description{
Make interactive bar charts utilizing \link[ggiraph]{ggiraph}.
}
