% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crosstab.R
\name{make_crosstab}
\alias{make_crosstab}
\title{Creates a formatted flat contingency table (\link[stats]{ftable}).}
\usage{
make_crosstab(tbl, col_level_threshold = 8, row_vars = NULL,
  col_vars = NULL, round_digits = 2)
}
\arguments{
\item{tbl}{A data.table returned by \link[NHTS.summarizer]{make_table}.}

\item{col_level_threshold}{Maximum amount of factor levels in the column position.}

\item{row_vars}{Factors to be represented in the row position.}

\item{col_vars}{Factors to be represented in the column position.}

\item{round_digits}{Number of digits to round to. Passed to digits parameter in \link[base]{round}.}
}
\description{
Creates a formatted flat contingency table (\link[stats]{ftable}).
}
