% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Create a map from geographically aggregated NHTS data}
\usage{
make_map(table1, geography = "state", table2)
}
\arguments{
\item{table1}{table returned from \link[summarizeNHTS]{make_table}}

\item{geography}{"state", "cbsa"}

\item{table2}{same as table1, but optional second table to show inside tooltip when geography is hovered with mouse}
}
\value{
ggiraph/htmlwidget class object
}
\description{
Create a map from geographically aggregated NHTS data
}
