% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Create a map from geographically aggregated NHTS data}
\usage{
make_map(tbl, tbl2, state_style = "normal", geo_layer = NULL,
  confidence = 0.95, ggiraph_options = list(), ...)
}
\arguments{
\item{tbl}{data.table returned from \link[summarizeNHTS]{summarize_data}, requires at least one group variable in tbl to be either HHSTATE or HH_CBSA (2009, 2017 only).}

\item{tbl2}{optional second table returned from \link[summarizeNHTS]{summarize_data}, requires same geography group variable as tbl. tbl2 gets passed to \link[summarizeNHTS]{make_chart} and output as an interactive tooltip over the matching tbl geography.}

\item{state_style}{either "normal" for typical state map boundaries or "tile" for tilemap style fixed-area boundaries.}

\item{geo_layer}{Optional geography layer specification. only required for derived/custom geography variables.}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.95. Set to NULL for standard error.}

\item{ggiraph_options}{Optional list of options to pass to \link[ggiraph]{ggiraph}.}

\item{...}{Optional formatting arguments. See \link[summarizeNHTS]{format_values}.}
}
\value{
ggiraph/htmlwidget class object
}
\description{
Create a map from geographically aggregated NHTS data
}
