% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create weighted aggregate tables using NHTS data.}
\usage{
make_table(data, agg, agg_var = NULL, factors = NULL, subset = TRUE,
  label = FALSE, variance = "se", jk_coeff = 99/100)
}
\arguments{
\item{data}{data.table object containing relavent NHTS variables and weights}

\item{agg}{Aggregate function label ("household_count", "person_count", "trip_count", 
"sum", "avg", "household_trip_rate", or "person_trip_rate")}

\item{agg_var}{Variable name to aggregate over. Only relavent when agg is "avg" or "sum"}

\item{factors}{Character element or vector of variable names to group by}

\item{subset}{A Pre-aggregation subset condition}

\item{label}{logical. Use labels for table output?}

\item{variance}{Variance calculation to be used. Either "se" for Standard Error or "moe" for Margin of Error}

\item{jk_coeff}{Jacknife coefficient for standard error calculations}
}
\value{
data.table object aggregated by input specifications
}
\description{
Create weighted aggregate tables using NHTS data.
}
