% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make HTML Table}
\usage{
make_table(tbl, title = "", ...)
}
\arguments{
\item{tbl}{An data.table object returned by \link[summarizeNHTS]{summarize_data}}

\item{title}{Title of the table}

\item{...}{Other arguments passed to \link[summarizeNHTS]{make_crosstab}. 
Includes formatting arguments. See \link[summarizeNHTS]{format_values}.}
}
\description{
Make formatted HTML tables utilizing \link[htmlTable]{htmlTable}
}
