% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create weighted aggregate tables using NHTS data.}
\usage{
make_table(data, agg, agg_var = NULL, factors = NULL, subset = TRUE,
<<<<<<< HEAD
  label = FALSE, prop = FALSE, prop_by = NULL, exclude_missing = FALSE)
=======
  label = TRUE, prop = FALSE, prop_by = NULL)
>>>>>>> c54a9e5fc1f919f957c261c856bb664f8377485a
}
\arguments{
\item{data}{Object returned by \link[summarizeNHTS]{read_nhts_data}}

\item{agg}{Aggregate function label. Either "household_count", "person_count", "trip_count", 
"sum", "avg", "household_trip_rate", or "person_trip_rate"}

\item{agg_var}{Character string specifying a numeric variable over which to aggregate. Only relavent when agg is "avg" or "sum"}

\item{factors}{Character vector of one or more variable names to group by}

\item{subset}{Character string containing a pre-aggregation subset condition using \link[data.table]{data.table} syntax}

\item{label}{logical. Use labels for table output?}

\item{prop}{logical. Use proportions for count aggregates?}

\item{prop_by}{Character vector of one or more variable names by which to group proportions}
}
\value{
data.table object aggregated by input specifications
}
\description{
Create weighted aggregate tables using NHTS data.
}
