% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read HTS data.}
\usage{
read_data(dataset, select = select_all(dataset), csv_path = getwd())
}
\arguments{
\item{dataset}{The study year of the dataset. Currently supports "2001" and "2009".}

\item{select}{A character vector of NHTS variable names to select for analysis. Defaults to all variables in the codebook.}

\item{csv_path}{The parent directory of "/csv/dataset/". Defaults to working directory.}
}
\value{
read_data returns an object of class "HTS.data". It contains the data files and weights necessary for
querying the NHTS dataset, used primarily by \link[summarizeNHTS]{summarize_data}.

The "HTS.data" obeject is essentially a list of data.tables broken up by data and weights.

Accessing the data:
\itemize{
  \item \strong{household} Household data file
  \item \strong{person} Person data file
  \item \strong{trip} item Trip data file
  \item \strong{vehicle} Vehicle data file
}

Accessing the weights:
\itemize{
  \item \strong{household} Household weight file. used for weighting household and vehicle data.
  \item \strong{person} Person weight file. Also includes trip weights at the person level.
}
}
\description{
Read and merge NHTS data for analysis.
}
\details{
\code{read_data} is a wrapper for reading in variables from the correct csvs
and merging the corresponding tables on correct keys. \link[data.table]{fread} and
\link[data.table]{merge} are used for performance benefits.
}
