% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhts_data.R
\name{read_nhts_data}
\alias{read_nhts_data}
\title{Read NHTS data.}
\usage{
read_nhts_data(dataset, select, csv_path = getwd())
}
\arguments{
\item{dataset}{The study year of the dataset to read Currently only '2009'.}

\item{select}{A character vector of NHTS variable names to select for analysis. 
See \link[NHTS.summarizer]{nhts_2009} for more info.}

\item{csv_path}{The parent directory of "/csv/dataset/". Defaults to working directory.}
}
\description{
Read and merge NHTS data for analysis.
}
\details{
\code{read_nhts_data} is a wrapper for reading in variables from the correct csvs
and merging the corresponding tables on correct keys. \link[data.table]{fread} and
\link[data.table]{merge.data.table} are used for performance benefits. The data user, 
however, must still understand the relationships between the variables, weights, tables, 
etc. to ensure that analyses are being performed correctly. For example, if a trip level
variable is selected, this function assumes you will be performing a trip-level analysis 
and read in trip-level weights. It is recommended to make different \code{read_nhts_data}
calls for different "level" analyses (household, person, trip, vehicle).
}
